requirements.txt
Django
httplib2
django-piston
